Refactor: 편집 버튼 폴더 위치 이동
